requirements.txt
numpy==1.26.3
paho-mqtt==2.0.0
pyserial==3.5
PyYAML==6.0.1
pyinstaller==6.3.0
requests==2.31.0
pydantic==2.5.2